Change weight dtype to float

+++ apis/models.py
@@ -9,7 +9,7 @@ class Bin(models.Model):
     color = models.CharField(max_length=50, null=True, blank=True)
     serial_number = models.CharField(max_length=50, null=False, blank=False)
     current_level = models.IntegerField(default=100)
-    current_weight = models.IntegerField(default=0)
+    current_weight =models.FloatField(default=0.0)
     bin_height = models.IntegerField(default=100)
     pickups = models.ManyToManyField('Pickup', related_name='bin', blank=True)
     latitude = models.FloatField(null=True, blank=True)
